fix: address code quality issues — memory leaks, file decomposition, duplication
Critical fixes:
- Bounded Polygon caches with max-size eviction and expired entry cleanup
- Rate limiter now deletes empty IP entries to prevent unbounded memory growth
- Alpaca broker API reuses a single httpx client with connection pooling

Backend decomposition:
- Extract submit_order() into order_service.py (halt checks, price fetch, risk check)
- Extract scan_signals() into signal_service.py (DB records, market data, signal gen)
- Split agents server.ts (543 lines) into 6 focused route modules

Frontend decomposition:
- orders/page.tsx: 1,080 → 298 lines (8 extracted components)
- regime/page.tsx: 764 → 206 lines (6 extracted components)
- autonomy/page.tsx: 758 → 35 lines (7 extracted components)

Code duplication:
- Created useMutationWithInvalidation generic hook wrapper
- Moved 7 mutation files from hooks/queries/ to hooks/mutations/
- Created mutations/index.ts barrel export

Strategy constants:
- Extracted all magic numbers into named constants with documentation
- Updated momentum.py and mean_reversion.py to use named constants

All 1,119 JS/TS tests pass. Lint clean across all 3 packages.
