+++ apps/agents/src/routes/index.ts
@@ -1,0 +1,6 @@
+export { healthRouter } from './health.js';
+export { orchestratorRouter } from './orchestrator.js';
+export { recommendationsRouter } from './recommendations.js';
+export { alertsRouter } from './alerts.js';
+export { agentRunsRouter } from './agent-runs.js';
+export { researchRouter } from './research.js';
