+++ apps/web/src/app/(dashboard)/orders/page.tsx
@@ -1,575 +1,25 @@
 'use client';
 
 import { useState, useMemo, useCallback } from 'react';
-import Link from 'next/link';
-import {
-  ArrowUpDown,
-  CheckCircle,
-  Shield,
-  ChevronDown,
-  ChevronUp,
-  Search,
-  Clock,
-  Activity,
-  DollarSign,
-  BarChart3,
-  ExternalLink,
-  FileText,
-  XCircle,
-  TrendingUp,
-  TrendingDown,
-} from 'lucide-react';
+import { ArrowUpDown, Activity } from 'lucide-react';
 import { Card, CardContent } from '@/components/ui/card';
-import {
-  Table,
-  TableHeader,
-  TableBody,
-  TableRow,
-  TableHead,
-  TableCell,
-  SortableTableHead,
-} from '@/components/ui/table';
-import type { SortState } from '@/components/ui/table';
-import { Select } from '@/components/ui/select';
-import { Input } from '@/components/ui/input';
 import { ErrorState } from '@/components/ui/error-state';
 import { ErrorBoundary } from '@/components/error-boundary';
 import { Spinner } from '@/components/ui/spinner';
 import { useFillsQuery, useRiskEvaluationsQuery, useOrderHistoryQuery } from '@/hooks/queries';
-import type { Fill, RiskEvaluation, RiskCheck } from '@sentinel/shared';
+import type { Fill, RiskEvaluation } from '@sentinel/shared';
 import type { OrderHistoryEntry } from '@/hooks/queries/use-order-history-query';
-import { cn } from '@/lib/utils';
-import { orderStatusColors, DEFAULT_ORDER_STYLE, sideColors } from '@/lib/status-colors';
+import type { SortState } from '@/components/ui/table';
 import { PAGE_SIZE_ORDERS } from '@/lib/constants';
-
-// ─── Types ──────────────────────────────────────────────────────────────
-
-type TimelineEntryType = 'fill' | 'risk' | 'order';
-
-interface TimelineEntry {
-  id: string;
-  type: TimelineEntryType;
-  timestamp: string;
-  data: Fill | RiskEvaluation | OrderHistoryEntry;
-}
-
-type DateRange = 'today' | 'week' | 'month' | 'all';
-type TypeFilter = 'all' | 'fills' | 'risk' | 'orders';
-
-// ─── Helpers ────────────────────────────────────────────────────────────
-
-function getDateRangeFrom(range: DateRange): string | undefined {
-  if (range === 'all') return undefined;
-  const now = new Date();
-  if (range === 'today') {
-    now.setHours(0, 0, 0, 0);
-  } else if (range === 'week') {
-    now.setDate(now.getDate() - 7);
-  } else if (range === 'month') {
-    now.setMonth(now.getMonth() - 1);
-  }
-  return now.toISOString();
-}
-
-function formatTimestamp(iso: string): string {
-  const d = new Date(iso);
-  return `${d.toLocaleDateString()} ${d.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit', second: '2-digit' })}`;
-}
-
-function formatCurrency(val: number): string {
-  return `$${val.toLocaleString(undefined, { minimumFractionDigits: 2, maximumFractionDigits: 2 })}`;
-}
-
-// ─── Type Metadata ──────────────────────────────────────────────────────
-
-const ENTRY_META: Record<TimelineEntryType, { label: string; icon: typeof CheckCircle }> = {
-  fill: { label: 'Fill', icon: CheckCircle },
-  risk: { label: 'Risk Eval', icon: Shield },
-  order: { label: 'Order', icon: FileText },
-};
-
-const ORDER_STATUS_STYLES = orderStatusColors;
-
-function getEntryBadgeColor(entry: TimelineEntry): string {
-  if (entry.type === 'fill') {
-    return 'bg-green-500/15 text-green-400';
-  }
-  if (entry.type === 'order') {
-    const order = entry.data as OrderHistoryEntry;
-    const style = ORDER_STATUS_STYLES[order.status] ?? DEFAULT_ORDER_STYLE;
-    return `${style.bg} ${style.text}`;
-  }
-  const risk = entry.data as RiskEvaluation;
-  return risk.allowed ? 'bg-orange-500/15 text-orange-400' : 'bg-red-500/15 text-red-400';
-}
-
-// ─── Stats Row ──────────────────────────────────────────────────────────
-
-function StatsRow({
-  fills,
-  riskEvals,
-  orders,
-  isLoading,
-}: {
-  fills: Fill[];
-  riskEvals: RiskEvaluation[];
-  orders: OrderHistoryEntry[];
-  isLoading: boolean;
-}) {
-  const stats = useMemo(() => {
-    const totalFills = fills.length;
-    const totalRiskEvals = riskEvals.length;
-    const totalOrders = orders.length;
-    const totalEntries = totalFills + totalRiskEvals + totalOrders;
-
-    const totalCommission = fills.reduce((sum, f) => sum + (f.commission ?? 0), 0);
-
-    const slippageValues = fills.filter((f) => f.slippage != null).map((f) => f.slippage as number);
-    const avgSlippage =
-      slippageValues.length > 0
-        ? slippageValues.reduce((a, b) => a + b, 0) / slippageValues.length
-        : null;
-
-    const fillRate =
-      totalOrders > 0
-        ? ((orders.filter((o) => o.status === 'filled').length / totalOrders) * 100).toFixed(1)
-        : '—';
-
-    return { totalEntries, totalFills, totalOrders, fillRate, totalCommission, avgSlippage };
-  }, [fills, riskEvals, orders]);
-
-  if (isLoading) {
-    return (
-      <div className="stagger-grid grid grid-cols-2 gap-2 sm:gap-3 lg:grid-cols-4">
-        {Array.from({ length: 4 }).map((_, i) => (
-          <Card key={i} className="border-zinc-800 bg-zinc-900/50">
-            <CardContent className="flex items-center justify-center p-4">
-              <Spinner size="sm" />
-            </CardContent>
-          </Card>
-        ))}
-      </div>
-    );
-  }
-
-  const cells = [
-    {
-      label: 'Total Events',
-      value: String(stats.totalEntries),
-      icon: Activity,
-    },
-    {
-      label: 'Fill Rate',
-      value: `${stats.fillRate}%`,
-      icon: BarChart3,
-    },
-    {
-      label: 'Total Commission',
-      value: formatCurrency(stats.totalCommission),
-      icon: DollarSign,
-    },
-    {
-      label: 'Avg Slippage',
-      value: stats.avgSlippage != null ? `${stats.avgSlippage.toFixed(4)}` : '—',
-      icon: ArrowUpDown,
-    },
-  ];
-
-  return (
-    <div className="stagger-grid grid grid-cols-2 gap-2 sm:gap-3 lg:grid-cols-4">
-      {cells.map((c) => {
-        const Icon = c.icon;
-        return (
-          <Card key={c.label} className="border-zinc-800 bg-zinc-900/50">
-            <CardContent className="p-4">
-              <div className="flex items-center gap-1.5">
-                <Icon className="h-3.5 w-3.5 text-zinc-500" />
-                <p className="text-xs text-zinc-500">{c.label}</p>
-              </div>
-              <p className="mt-1 text-data-primary text-zinc-100">{c.value}</p>
-            </CardContent>
-          </Card>
-        );
-      })}
-    </div>
-  );
-}
-
-// ─── Risk Check Row ─────────────────────────────────────────────────────
-
-function RiskCheckRow({ check }: { check: RiskCheck }) {
-  return (
-    <div className="flex items-center gap-2 text-xs">
-      <span
-        className={`h-1.5 w-1.5 rounded-full ${check.passed ? 'bg-green-400' : 'bg-red-400'}`}
-      />
-      <span className="text-zinc-300">{check.name}</span>
-      {check.actual != null && check.limit != null && (
-        <span className="text-zinc-500">
-          ({check.actual} / {check.limit})
-        </span>
-      )}
-      {check.message && (
-        <span className="text-zinc-600 truncate max-w-[150px] sm:max-w-xs">{check.message}</span>
-      )}
-    </div>
-  );
-}
-
-// ─── Fill Detail ────────────────────────────────────────────────────────
-
-function FillDetail({ fill }: { fill: Fill }) {
-  return (
-    <div className="space-y-2">
-      <div className="grid grid-cols-2 gap-x-3 gap-y-1 text-xs sm:grid-cols-3 sm:gap-x-6">
-        <div>
-          <span className="text-zinc-500">Order ID</span>
-          <p className="text-zinc-300 font-mono text-[11px] truncate">{fill.order_id}</p>
-        </div>
-        <div>
-          <span className="text-zinc-500">Fill Price</span>
-          <p className="text-zinc-300">{formatCurrency(fill.fill_price)}</p>
-        </div>
-        <div>
-          <span className="text-zinc-500">Fill Qty</span>
-          <p className="text-zinc-300">{fill.fill_qty}</p>
-        </div>
-        <div>
-          <span className="text-zinc-500">Commission</span>
-          <p className="text-zinc-300">
-            {fill.commission != null ? formatCurrency(fill.commission) : '—'}
-          </p>
-        </div>
-        <div>
-          <span className="text-zinc-500">Slippage</span>
-          <p className="text-zinc-300">{fill.slippage != null ? fill.slippage.toFixed(4) : '—'}</p>
-        </div>
-        <div>
-          <span className="text-zinc-500">Venue</span>
-          <p className="text-zinc-300">{(fill as Fill & { venue?: string | null }).venue ?? '—'}</p>
-        </div>
-      </div>
-      {(fill as Fill & { broker_fill_id?: string | null }).broker_fill_id && (
-        <div className="text-xs">
-          <span className="text-zinc-500">Broker Fill ID: </span>
-          <span className="text-zinc-400 font-mono">
-            {(fill as Fill & { broker_fill_id?: string | null }).broker_fill_id}
-          </span>
-        </div>
-      )}
-    </div>
-  );
-}
-
-// ─── Risk Eval Detail ───────────────────────────────────────────────────
-
-function RiskEvalDetail({ evaluation }: { evaluation: RiskEvaluation }) {
-  return (
-    <div className="space-y-2">
-      <div className="grid grid-cols-2 gap-x-3 gap-y-1 text-xs sm:grid-cols-3 sm:gap-x-6">
-        <div>
-          <span className="text-zinc-500">Recommendation</span>
-          <div className="flex items-center gap-1">
-            <p className="text-zinc-300 font-mono text-[11px] truncate">
-              {evaluation.recommendation_id}
-            </p>
-            <Link
-              href={`/recommendations/${evaluation.recommendation_id}`}
-              className="text-blue-400 hover:text-blue-300 transition-colors"
-            >
-              <ExternalLink className="h-3 w-3" />
-            </Link>
-          </div>
-        </div>
-        <div>
-          <span className="text-zinc-500">Policy Version</span>
-          <p className="text-zinc-300">{evaluation.policy_version ?? '—'}</p>
-        </div>
-        <div>
-          <span className="text-zinc-500">Decision</span>
-          <p className={evaluation.allowed ? 'text-green-400' : 'text-red-400'}>
-            {evaluation.allowed ? 'Allowed' : 'Blocked'}
-          </p>
-        </div>
-        {evaluation.adjusted_quantity != null && (
-          <div>
-            <span className="text-zinc-500">Adjusted Qty</span>
-            <p className="text-zinc-300">{evaluation.adjusted_quantity}</p>
-          </div>
-        )}
-        {(evaluation as RiskEvaluation & { original_quantity?: number | null }).original_quantity !=
-          null && (
-          <div>
-            <span className="text-zinc-500">Original Qty</span>
-            <p className="text-zinc-300">
-              {
-                (
-                  evaluation as RiskEvaluation & {
-                    original_quantity?: number | null;
-                  }
-                ).original_quantity
-              }
-            </p>
-          </div>
-        )}
-      </div>
-
-      {evaluation.reason && (
-        <div className="text-xs">
-          <span className="text-zinc-500">Reason: </span>
-          <span className="text-zinc-300">{evaluation.reason}</span>
-        </div>
-      )}
-
-      {evaluation.checks_performed.length > 0 && (
-        <div>
-          <p className="mb-1 text-xs font-medium text-zinc-500">
-            Checks ({evaluation.checks_performed.length})
-          </p>
-          <div className="space-y-1">
-            {evaluation.checks_performed.map((check, idx) => (
-              <RiskCheckRow key={`${check.name}-${idx}`} check={check} />
-            ))}
-          </div>
-        </div>
-      )}
-    </div>
-  );
-}
-
-// ─── Execution Quality ──────────────────────────────────────────────────
-
-function ExecutionQuality({ order }: { order: OrderHistoryEntry }) {
-  if (order.fill_price == null || order.status !== 'filled') return null;
-
-  return (
-    <div className="mt-2 rounded border border-zinc-800 bg-zinc-950/50 p-2">
-      <p className="text-[10px] font-medium text-zinc-500 uppercase tracking-wider mb-1">
-        Execution Quality
-      </p>
-      <div className="flex flex-wrap gap-3 text-xs">
-        <div>
-          <span className="text-zinc-500">Fill Price: </span>
-          <span className="text-zinc-200 font-medium">{formatCurrency(order.fill_price)}</span>
-        </div>
-        {order.order_type !== 'market' && (
-          <div>
-            <span className="text-zinc-500">
-              {order.order_type === 'limit' ? 'Limit' : 'Request'} Price:{' '}
-            </span>
-            <span className="text-zinc-200">
-              {order.fill_price ? formatCurrency(order.fill_price) : '—'}
-            </span>
-          </div>
-        )}
-        <div>
-          <span className="text-zinc-500">Fill Rate: </span>
-          <span
-            className={cn(
-              'font-medium',
-              order.filled_qty >= order.qty ? 'text-green-400' : 'text-amber-400',
-            )}
-          >
-            {order.qty > 0 ? ((order.filled_qty / order.qty) * 100).toFixed(0) : 0}%
-          </span>
-          <span className="text-zinc-600 ml-1">
-            ({order.filled_qty}/{order.qty})
-          </span>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-// ─── Order Detail ───────────────────────────────────────────────────────
-
-const ORDER_STATUS_STEPS = ['pending', 'submitted', 'partial', 'filled'] as const;
-const ORDER_TERMINAL = ['cancelled', 'rejected'] as const;
-
-function OrderLifecycleBar({ status }: { status: string }) {
-  const isTerminal = (ORDER_TERMINAL as readonly string[]).includes(status);
-  const currentIdx = isTerminal
-    ? -1
-    : ORDER_STATUS_STEPS.indexOf(status as (typeof ORDER_STATUS_STEPS)[number]);
-
-  return (
-    <div className="flex items-center gap-1 mt-2">
-      {ORDER_STATUS_STEPS.map((step, idx) => {
-        const isActive = !isTerminal && idx <= currentIdx;
-        const isCurrent = !isTerminal && idx === currentIdx;
-        return (
-          <div key={step} className="flex items-center gap-1">
-            <div
-              className={cn(
-                'h-1.5 rounded-full transition-all',
-                idx === 0 ? 'w-4' : 'w-8',
-                isActive ? (isCurrent ? 'bg-green-400' : 'bg-green-400/50') : 'bg-zinc-800',
-              )}
-            />
-            {idx < ORDER_STATUS_STEPS.length - 1 && (
-              <div
-                className={cn('w-1 h-1 rounded-full', isActive ? 'bg-green-400/30' : 'bg-zinc-800')}
-              />
-            )}
-          </div>
-        );
-      })}
-      {isTerminal && (
-        <div className="flex items-center gap-1 ml-1">
-          <XCircle className="h-3 w-3 text-red-400" />
-          <span className="text-[10px] text-red-400 font-medium uppercase">{status}</span>
-        </div>
-      )}
-    </div>
-  );
-}
-
-function OrderDetail({ order }: { order: OrderHistoryEntry }) {
-  return (
-    <div className="space-y-2">
-      {/* Lifecycle bar */}
-      <div>
-        <p className="text-[10px] font-medium text-zinc-500 uppercase tracking-wider">
-          Order Lifecycle
-        </p>
-        <OrderLifecycleBar status={order.status} />
-        <div className="flex gap-2 sm:gap-3 mt-1 text-[10px] text-zinc-600">
-          {ORDER_STATUS_STEPS.map((s) => (
-            <span key={s}>{s}</span>
-          ))}
-        </div>
-      </div>
-
-      <div className="grid grid-cols-2 gap-x-3 gap-y-1 text-xs sm:grid-cols-3 sm:gap-x-6">
-        <div>
-          <span className="text-zinc-500">Order ID</span>
-          <p className="text-zinc-300 font-mono text-[11px] truncate">{order.order_id}</p>
-        </div>
-        <div>
-          <span className="text-zinc-500">Symbol</span>
-          <p className="text-zinc-300 font-semibold">{order.symbol}</p>
-        </div>
-        <div>
-          <span className="text-zinc-500">Side</span>
-          <p
-            className={cn(
-              'inline-flex rounded-full px-2 py-0.5 text-xs font-medium',
-              sideColors[order.side] ?? 'text-zinc-400',
-            )}
-          >
-            {order.side.toUpperCase()}
-          </p>
-        </div>
-        <div>
-          <span className="text-zinc-500">Type</span>
-          <p className="text-zinc-300">{order.order_type}</p>
-        </div>
-        <div>
-          <span className="text-zinc-500">Quantity</span>
-          <p className="text-zinc-300">{order.qty}</p>
-        </div>
-        <div>
-          <span className="text-zinc-500">Filled Qty</span>
-          <p className="text-zinc-300">{order.filled_qty}</p>
-        </div>
-        {order.fill_price != null && (
-          <div>
-            <span className="text-zinc-500">Fill Price</span>
-            <p className="text-zinc-300">{formatCurrency(order.fill_price)}</p>
-          </div>
-        )}
-        {order.submitted_at && (
-          <div>
-            <span className="text-zinc-500">Submitted</span>
-            <p className="text-zinc-300">{formatTimestamp(order.submitted_at)}</p>
-          </div>
-        )}
-        {order.filled_at && (
-          <div>
-            <span className="text-zinc-500">Filled</span>
-            <p className="text-zinc-300">{formatTimestamp(order.filled_at)}</p>
-          </div>
-        )}
-      </div>
-
-      {order.risk_note && (
-        <div className="text-xs">
-          <span className="text-zinc-500">Risk Note: </span>
-          <span className="text-zinc-300">{order.risk_note}</span>
-        </div>
-      )}
-
-      <ExecutionQuality order={order} />
-    </div>
-  );
-}
-
-// ─── Date Range Buttons ─────────────────────────────────────────────────
-
-const DATE_RANGE_OPTIONS: { value: DateRange; label: string }[] = [
-  { value: 'today', label: 'Today' },
-  { value: 'week', label: 'Week' },
-  { value: 'month', label: 'Month' },
-  { value: 'all', label: 'All' },
-];
-
-const TYPE_FILTER_OPTIONS: { value: TypeFilter; label: string }[] = [
-  { value: 'all', label: 'All Types' },
-  { value: 'orders', label: 'Orders' },
-  { value: 'fills', label: 'Fills' },
-  { value: 'risk', label: 'Risk Evals' },
-];
-
-// ─── Sort Helpers ───────────────────────────────────────────────────────
-
-function getEntrySymbol(entry: TimelineEntry): string {
-  if (entry.type === 'order') return (entry.data as OrderHistoryEntry).symbol;
-  if (entry.type === 'fill') return (entry.data as Fill).order_id.slice(0, 8);
-  return (entry.data as RiskEvaluation).recommendation_id.slice(0, 8);
-}
-
-function getEntryStatus(entry: TimelineEntry): string {
-  if (entry.type === 'fill') return 'filled';
-  if (entry.type === 'order') return (entry.data as OrderHistoryEntry).status;
-  return (entry.data as RiskEvaluation).allowed ? 'allowed' : 'blocked';
-}
-
-function getEntrySide(entry: TimelineEntry): string | null {
-  if (entry.type === 'order') return (entry.data as OrderHistoryEntry).side;
-  return null;
-}
-
-function getEntrySummary(entry: TimelineEntry): { primary: string; secondary: string | null } {
-  if (entry.type === 'fill') {
-    const fill = entry.data as Fill;
-    return {
-      primary: `${fill.fill_qty} @ ${formatCurrency(fill.fill_price)}`,
-      secondary: fill.order_id ? `Order ${fill.order_id.slice(0, 8)}…` : null,
-    };
-  }
-  if (entry.type === 'order') {
-    const order = entry.data as OrderHistoryEntry;
-    const sideLabel = order.side === 'buy' ? 'BUY' : 'SELL';
-    return {
-      primary: `${sideLabel} ${order.symbol} × ${order.qty}`,
-      secondary:
-        order.fill_price != null
-          ? `Filled @ ${formatCurrency(order.fill_price)}`
-          : order.status === 'cancelled'
-            ? 'Cancelled'
-            : order.status === 'rejected'
-              ? 'Rejected'
-              : null,
-    };
-  }
-  const risk = entry.data as RiskEvaluation;
-  return {
-    primary: risk.allowed
-      ? `Approved${risk.adjusted_quantity != null ? ` (qty: ${risk.adjusted_quantity})` : ''}`
-      : 'Blocked',
-    secondary: risk.reason,
-  };
-}
+import {
+  StatsRow,
+  OrderFilters,
+  OrderTimelineTable,
+  getDateRangeFrom,
+  getEntrySymbol,
+  getEntryStatus,
+} from '@/components/orders';
+import type { DateRange, TypeFilter, TimelineEntry } from '@/components/orders';
 
 // ─── Main Page ──────────────────────────────────────────────────────────
 
@@ -735,6 +185,11 @@ export default function OrdersPage() {
     setPage(0);
   }, []);
 
+  const handleSymbolSearch = useCallback((value: string) => {
+    setSymbolSearch(value);
+    setPage(0);
+  }, []);
+
   const handleClearFilters = useCallback(() => {
     setDateRange('all');
     setTypeFilter('all');
@@ -760,83 +215,17 @@ export default function OrdersPage() {
         <StatsRow fills={fills} riskEvals={riskEvals} orders={orders} isLoading={isLoading} />
 
         {/* Filters */}
-        <div
-          className="flex flex-col gap-3 sm:flex-row sm:flex-wrap sm:items-center"
-          role="search"
-          aria-label="Filter orders"
-        >
-          <div className="flex items-center gap-2 overflow-x-auto">
-            {/* Date range */}
-            <div
-              className="flex shrink-0 rounded-md border border-zinc-700 overflow-hidden"
-              role="group"
-              aria-label="Date range filter"
-            >
-              {DATE_RANGE_OPTIONS.map((opt) => (
-                <button
-                  key={opt.value}
-                  onClick={() => handleDateRange(opt.value)}
-                  aria-label={`Filter by ${opt.label}`}
-                  aria-pressed={dateRange === opt.value}
-                  className={cn(
-                    'px-3 py-1.5 text-xs transition-colors',
-                    dateRange === opt.value
-                      ? 'bg-zinc-700 text-zinc-100 font-medium'
-                      : 'bg-zinc-900 text-zinc-400 hover:text-zinc-200 hover:bg-zinc-800',
-                  )}
-                >
-                  {opt.label}
-                </button>
-              ))}
-            </div>
-
-            {/* Type filter */}
-            <Select
-              value={typeFilter}
-              onChange={(e) => handleTypeFilter(e.target.value as TypeFilter)}
-              aria-label="Filter by event type"
-              className="shrink-0 w-auto"
-            >
-              {TYPE_FILTER_OPTIONS.map((opt) => (
-                <option key={opt.value} value={opt.value}>
-                  {opt.label}
-                </option>
-              ))}
-            </Select>
-          </div>
-
-          <div className="flex items-center gap-2">
-            {/* Symbol / ID search */}
-            <div className="relative flex-1 sm:flex-initial">
-              <Search className="absolute left-2.5 top-1/2 h-3.5 w-3.5 -translate-y-1/2 text-zinc-500 z-10" />
-              <Input
-                type="text"
-                placeholder="Search ID or reason…"
-                value={symbolSearch}
-                onChange={(e) => {
-                  setSymbolSearch(e.target.value);
-                  setPage(0);
-                }}
-                aria-label="Search orders by ID, symbol, or reason"
-                className="w-full pl-8 sm:w-48"
-              />
-            </div>
-
-            {hasFilters && (
-              <button
-                onClick={handleClearFilters}
-                aria-label="Clear all filters"
-                className="shrink-0 text-xs text-zinc-500 hover:text-zinc-300 transition-colors"
-              >
-                Clear
-              </button>
-            )}
-          </div>
-
-          <span className="text-xs text-zinc-600 sm:ml-auto">
-            {totalEntries} {totalEntries === 1 ? 'event' : 'events'}
-          </span>
-        </div>
+        <OrderFilters
+          dateRange={dateRange}
+          typeFilter={typeFilter}
+          symbolSearch={symbolSearch}
+          totalEntries={totalEntries}
+          hasFilters={hasFilters}
+          onDateRange={handleDateRange}
+          onTypeFilter={handleTypeFilter}
+          onSymbolSearch={handleSymbolSearch}
+          onClearFilters={handleClearFilters}
+        />
 
         {/* Timeline Table */}
         {isLoading && (
@@ -870,184 +259,13 @@ export default function OrdersPage() {
         )}
 
         {!isLoading && !isError && pagedEntries.length > 0 && (
-          <Card className="border-zinc-800 bg-zinc-900/50 overflow-hidden">
-            <Table aria-label="Orders and execution timeline">
-              <TableHeader>
-                <TableRow>
-                  <SortableTableHead column="type" sortState={sortState} onSort={handleSort}>
-                    Type
-                  </SortableTableHead>
-                  <SortableTableHead column="symbol" sortState={sortState} onSort={handleSort}>
-                    Symbol / ID
-                  </SortableTableHead>
-                  <TableHead>Side</TableHead>
-                  <SortableTableHead column="status" sortState={sortState} onSort={handleSort}>
-                    Status
-                  </SortableTableHead>
-                  <TableHead>Summary</TableHead>
-                  <SortableTableHead column="timestamp" sortState={sortState} onSort={handleSort}>
-                    Time
-                  </SortableTableHead>
-                  <TableHead className="w-10">
-                    <span className="sr-only">Actions</span>
-                  </TableHead>
-                </TableRow>
-              </TableHeader>
-              <TableBody>
-                {pagedEntries.map((entry) => {
-                  const meta = ENTRY_META[entry.type];
-                  const badgeColor = getEntryBadgeColor(entry);
-                  const Icon = meta.icon;
-                  const side = getEntrySide(entry);
-                  const status = getEntryStatus(entry);
-                  const statusStyle =
-                    entry.type === 'order'
-                      ? (ORDER_STATUS_STYLES[(entry.data as OrderHistoryEntry).status] ??
-                        DEFAULT_ORDER_STYLE)
-                      : null;
-                  const summary = getEntrySummary(entry);
-                  const isExpanded = expandedId === entry.id;
-
-                  return (
-                    <TableRow key={entry.id}>
-                      {/* Type badge */}
-                      <TableCell>
-                        <span
-                          className={cn(
-                            'inline-flex items-center gap-1 rounded-full px-2 py-0.5 text-xs font-medium',
-                            badgeColor,
-                          )}
-                        >
-                          <Icon className="h-3 w-3" />
-                          {meta.label}
-                        </span>
-                      </TableCell>
-
-                      {/* Symbol / ID */}
-                      <TableCell className="font-mono text-xs">
-                        {entry.type === 'order' ? (
-                          <span className="font-semibold text-zinc-100">
-                            {(entry.data as OrderHistoryEntry).symbol}
-                          </span>
-                        ) : entry.type === 'fill' ? (
-                          <span className="text-zinc-400 truncate max-w-[100px] inline-block">
-                            {(entry.data as Fill).order_id.slice(0, 8)}…
-                          </span>
-                        ) : (
-                          <Link
-                            href={`/recommendations/${(entry.data as RiskEvaluation).recommendation_id}`}
-                            className="text-blue-400 hover:text-blue-300 transition-colors inline-flex items-center gap-1"
-                          >
-                            {(entry.data as RiskEvaluation).recommendation_id.slice(0, 8)}…
-                            <ExternalLink className="h-3 w-3" />
-                          </Link>
-                        )}
-                      </TableCell>
-
-                      {/* Side */}
-                      <TableCell>
-                        {side ? (
-                          <span
-                            className={cn(
-                              'inline-flex items-center gap-1 rounded-full px-2 py-0.5 text-xs font-medium',
-                              sideColors[side] ?? '',
-                            )}
-                          >
-                            {side === 'buy' ? (
-                              <TrendingUp className="h-3 w-3" />
-                            ) : (
-                              <TrendingDown className="h-3 w-3" />
-                            )}
-                            {side.toUpperCase()}
-                          </span>
-                        ) : (
-                          <span className="text-zinc-600">—</span>
-                        )}
-                      </TableCell>
-
-                      {/* Status */}
-                      <TableCell>
-                        {statusStyle ? (
-                          <span
-                            className={cn(
-                              'inline-flex items-center gap-1 rounded-full px-2 py-0.5 text-xs font-medium',
-                              statusStyle.bg,
-                              statusStyle.text,
-                            )}
-                          >
-                            {status.charAt(0).toUpperCase() + status.slice(1)}
-                          </span>
-                        ) : (
-                          <span
-                            className={cn(
-                              'inline-flex items-center gap-1 rounded-full px-2 py-0.5 text-xs font-medium',
-                              badgeColor,
-                            )}
-                          >
-                            {status.charAt(0).toUpperCase() + status.slice(1)}
-                          </span>
-                        )}
-                      </TableCell>
-
-                      {/* Summary */}
-                      <TableCell className="max-w-[200px]">
-                        <span className="text-sm text-zinc-200">{summary.primary}</span>
-                        {summary.secondary && (
-                          <span className="block text-xs text-zinc-500 truncate">
-                            {summary.secondary}
-                          </span>
-                        )}
-                      </TableCell>
-
-                      {/* Time */}
-                      <TableCell className="text-xs text-zinc-500 whitespace-nowrap">
-                        <span className="inline-flex items-center gap-1">
-                          <Clock className="h-3 w-3" />
-                          <time dateTime={entry.timestamp}>{formatTimestamp(entry.timestamp)}</time>
-                        </span>
-                      </TableCell>
-
-                      {/* Expand */}
-                      <TableCell>
-                        <button
-                          onClick={() => setExpandedId(isExpanded ? null : entry.id)}
-                          aria-label={isExpanded ? 'Collapse details' : 'Expand details'}
-                          aria-expanded={isExpanded}
-                          className="text-zinc-500 hover:text-zinc-300 transition-colors"
-                        >
-                          {isExpanded ? (
-                            <ChevronUp className="h-4 w-4" />
-                          ) : (
-                            <ChevronDown className="h-4 w-4" />
-                          )}
-                        </button>
-                      </TableCell>
-                    </TableRow>
-                  );
-                })}
-              </TableBody>
-            </Table>
-
-            {/* Expanded detail rows */}
-            {expandedId && pagedEntries.some((e) => e.id === expandedId) && (
-              <div className="border-t border-zinc-800 bg-zinc-950/50 p-4">
-                {(() => {
-                  const entry = pagedEntries.find((e) => e.id === expandedId)!;
-                  return (
-                    <>
-                      {entry.type === 'fill' && <FillDetail fill={entry.data as Fill} />}
-                      {entry.type === 'risk' && (
-                        <RiskEvalDetail evaluation={entry.data as RiskEvaluation} />
-                      )}
-                      {entry.type === 'order' && (
-                        <OrderDetail order={entry.data as OrderHistoryEntry} />
-                      )}
-                    </>
-                  );
-                })()}
-              </div>
-            )}
-          </Card>
+          <OrderTimelineTable
+            pagedEntries={pagedEntries}
+            sortState={sortState}
+            expandedId={expandedId}
+            onSort={handleSort}
+            onToggleExpand={setExpandedId}
+          />
         )}
 
         {/* Pagination */}
