+++ apps/web/src/components/regime/index.ts
@@ -1,0 +1,12 @@
+export { CurrentRegimeCard } from './current-regime-card';
+export { RegimeSelector } from './regime-selector';
+export { PlaybookCard } from './playbook-card';
+export { CreatePlaybookDialog } from './create-playbook-dialog';
+export { RegimeTimeline } from './regime-timeline';
+export {
+  REGIMES,
+  STRATEGY_FAMILIES,
+  formatStrategy,
+  getRegimeMeta,
+  type RegimeMeta,
+} from './regime-constants';
