+++ apps/web/src/hooks/mutations/index.ts
@@ -1,0 +1,29 @@
+'use client';
+
+// ── Advisor ──────────────────────────────────────────────────────────
+export {
+  useCreatePreferenceMutation,
+  useUpdatePreferenceMutation,
+  useConfirmPreferenceMutation,
+  useDismissPreferenceMutation,
+  useDeletePreferenceMutation,
+} from './use-advisor-preference-mutations';
+export { useAdvisorProfileMutation } from './use-advisor-profile-mutations';
+export {
+  useCreateThreadMutation,
+  useUpdateThreadMutation,
+  useDeleteThreadMutation,
+} from './use-advisor-thread-mutations';
+
+// ── Trading ──────────────────────────────────────────────────────────
+export { useSubmitOrderMutation } from './use-submit-order-mutation';
+export { useApproveRecommendationMutation } from './use-approve-recommendation-mutation';
+export { useRejectRecommendationMutation } from './use-reject-recommendation-mutation';
+
+// ── Agents ───────────────────────────────────────────────────────────
+export { useTriggerCycleMutation } from './use-trigger-cycle-mutation';
+export { useHaltMutation, useResumeMutation } from './use-halt-resume-mutations';
+
+// ── Journal ──────────────────────────────────────────────────────────
+export { useCreateJournalEntryMutation } from './use-create-journal-entry-mutation';
+export { useGradeJournalMutation } from './use-grade-journal-mutation';
