+++ apps/web/src/hooks/use-mutation-with-invalidation.ts
@@ -1,0 +1,31 @@
+'use client';
+
+import {
+  useMutation,
+  useQueryClient,
+  type QueryKey,
+  type UseMutationOptions,
+} from '@tanstack/react-query';
+
+/**
+ * Generic mutation hook that invalidates specified query keys on success.
+ *
+ * Reduces boilerplate for the common pattern of:
+ *   useMutation({ mutationFn, onSuccess: () => queryClient.invalidateQueries(...) })
+ */
+export function useMutationWithInvalidation<TData, TVariables>(
+  mutationFn: (variables: TVariables) => Promise<TData>,
+  invalidateKeys: QueryKey[],
+  options?: Omit<UseMutationOptions<TData, Error, TVariables>, 'mutationFn' | 'onSuccess'>,
+) {
+  const queryClient = useQueryClient();
+  return useMutation({
+    mutationFn,
+    onSuccess: () => {
+      for (const key of invalidateKeys) {
+        queryClient.invalidateQueries({ queryKey: key });
+      }
+    },
+    ...options,
+  });
+}
